+++ tests/test_display/test_components/test_box_layout.py
@@ -1,5 +1,7 @@
 """Test the various box layout methods."""
 
+import cocos
+
 from shimmer.display.data_structures import Color
 from shimmer.display.widgets.button import ButtonDefinition, Button
 from shimmer.display.components.box_layout import (
@@ -31,8 +33,9 @@ def test_box_row(run_gui):
         buttons.append(btn)
 
     box_row = BoxRow(buttons, spacing=10)
-    assert box_row.rect.height == 100
     assert box_row.rect.width == 190
+    assert box_row.rect.height == 100
+    assert box_row.bounding_rect_of_children() == cocos.rect.Rect(0, 0, 190, 100)
 
     assert run_gui(test_box_row, box_row)
 
@@ -47,6 +50,7 @@ def test_box_column(run_gui):
     box_column = BoxColumn(buttons)
     assert box_column.rect.width == 100
     assert box_column.rect.height == 190
+    assert box_column.bounding_rect_of_children() == cocos.rect.Rect(0, 0, 100, 190)
 
     assert run_gui(test_box_column, box_column)
 
@@ -63,5 +67,6 @@ def test_build_rectangular_grid_horizontal(run_gui):
 
     assert box_grid.rect.width == 430
     assert box_grid.rect.height == 140
+    assert box_grid.bounding_rect_of_children() == cocos.rect.Rect(0, 0, 430, 140)
 
     assert run_gui(test_build_rectangular_grid_horizontal, box_grid)